Remove of duplicated method for frugal streaming
