+++ include/fdreadoutlibs/wib2/tpg/ProcessAVX2FIR.hpp
@@ -10,6 +10,7 @@
 #define READOUT_SRC_WIB2_TPG_PROCESSAVX2FIR_HPP_
 
 #include "FrameExpand.hpp"
+#include "UtilsAVX2.hpp"
 #include "ProcessingInfo.hpp"
 #include "TPGConstants_wib2.hpp"
 
@@ -17,64 +18,6 @@
 
 namespace swtpg_wib2 {
 
-inline void
-frugal_accum_update_avx2(__m256i& __restrict__ median,
-                         const __m256i s,
-                         __m256i& __restrict__ accum,
-                         const int16_t acclimit,
-                         const __m256i mask) __attribute__((always_inline));
-
-inline void
-frugal_accum_update_avx2(__m256i& __restrict__ median,
-                         const __m256i s,
-                         __m256i& __restrict__ accum,
-                         const int16_t acclimit,
-                         const __m256i mask)
-{
-  // if the sample is greater than the median, add one to the accumulator
-  // if the sample is less than the median, subtract one from the accumulator.
-
-  // For reasons that I don't understand, there's no cmplt
-  // for "compare less-than", so we have to compare greater,
-  // compare equal, and take everything else to be compared
-  // less-then
-  __m256i is_gt = _mm256_cmpgt_epi16(s, median);
-  __m256i is_eq = _mm256_cmpeq_epi16(s, median);
-
-  __m256i to_add = _mm256_set1_epi16(-1);
-  // Really want an epi16 version of this, but the cmpgt and
-  // cmplt functions set their epi16 parts to 0xff or 0x0,
-  // so treating everything as epi8 works the same
-  to_add = _mm256_blendv_epi8(to_add, _mm256_set1_epi16(1), is_gt);
-  to_add = _mm256_blendv_epi8(to_add, _mm256_set1_epi16(0), is_eq);
-
-  // Don't add anything to the channels which are masked out
-  to_add = _mm256_and_si256(to_add, mask);
-
-  accum = _mm256_add_epi16(accum, to_add);
-
-  // if the accumulator is >10, add one to the median and
-  // set the accumulator to zero. if the accumulator is
-  // <-10, subtract one from the median and set the
-  // accumulator to zero
-  is_gt = _mm256_cmpgt_epi16(accum, _mm256_set1_epi16(acclimit));
-  __m256i is_lt =
-    _mm256_cmpgt_epi16(_mm256_sign_epi16(accum, _mm256_set1_epi16(-1 * acclimit)), _mm256_set1_epi16(acclimit));
-
-  to_add = _mm256_setzero_si256();
-  to_add = _mm256_blendv_epi8(to_add, _mm256_set1_epi16(1), is_gt);
-  to_add = _mm256_blendv_epi8(to_add, _mm256_set1_epi16(-1), is_lt);
-
-  // Don't add anything to the channels which are masked out
-  to_add = _mm256_and_si256(to_add, mask);
-
-  median = _mm256_adds_epi16(median, to_add);
-
-  // Reset the unmasked channels that were >10 or <-10 to zero, leaving the others unchanged
-  __m256i need_reset = _mm256_or_si256(is_lt, is_gt);
-  need_reset = _mm256_and_si256(need_reset, mask);
-  accum = _mm256_blendv_epi8(accum, _mm256_setzero_si256(), need_reset);
-}
 
 template<size_t NREGISTERS>
 inline void
@@ -173,13 +116,13 @@ process_window_avx2(ProcessingInfo<NREGISTERS>& info)
       __m256i is_lt = _mm256_xor_si256(gt_or_eq, _mm256_set1_epi16(0xffff));
 #pragma GCC diagnostic pop
       // Update the 25th percentile in the channels that are below the median
-      frugal_accum_update_avx2(quantile25, s, accum25, 10, is_lt);
+      swtpg_wib2::frugal_accum_update_avx2(quantile25, s, accum25, 10, is_lt);
       // Update the 75th percentile in the channels that are above the median
-      frugal_accum_update_avx2(quantile75, s, accum75, 10, is_gt);
+      swtpg_wib2::frugal_accum_update_avx2(quantile75, s, accum75, 10, is_gt);
       // Update the median itself in all channels
 #pragma GCC diagnostic push
 #pragma GCC diagnostic ignored "-Woverflow"
-      frugal_accum_update_avx2(median, s, accum, 10, _mm256_set1_epi16(0xffff));
+      swtpg_wib2::frugal_accum_update_avx2(median, s, accum, 10, _mm256_set1_epi16(0xffff));
 #pragma GCC diagnostic pop
       // Actually subtract the pedestal
       s = _mm256_sub_epi16(s, median);
